Split gprc_server into grpc/mod.rs and grpc/tests.rs
